Use proquint as the format for device name.
The device name is meant to be a short, easy to remember and pronounce string.
The purpose of the name is to distinguish a device in a set of devices. This
patch implements device names using proquint.

We save the name in the sqlite database so that it won't change on every
restart. We also put the name into the device's certificate common name. If
there is an existing certificate and if its name does not match the device
name, re-generate the certificate.